requirements.txt
h5py==3.10.0
matplotlib==3.8.4
numpy==1.26.4
Pillow==10.0.0
scipy==1.11.1
torch==2.0.1
torchvision==0.15.2